requirements.txt
accelerate==0.30.1
deepspeed==0.14.4
einops==0.8.0
transformers==4.43.3
huggingface-hub==0.24.5
optimum-quanto
datasets
omegaconf
diffusers
sentencepiece
opencv-python
matplotlib
onnxruntime
torchvision
timm
python-dotenv
wandb

git+https://github.com/openai/CLIP.git
